+++ Source/Ariadne/Ariadne.swift
@@ -45,11 +45,11 @@ public protocol RootViewProvider {
 }
 
 /// Type, that is capable of performing route from current screen to another one.
-/// One example of such type is Route<ViewBuilder,ViewTransition> type, that includes necessary builder to build next visible view, and transition object, that will perform a transition.
+/// One example of such type is Route<ViewControllerBuilder,ViewTransition> type, that includes necessary builder to build next visible view, and transition object, that will perform a transition.
 public protocol Routable {
 
     /// Type, responsible for building a view, that is needed for routing.
-    associatedtype Builder: ViewBuilder
+    associatedtype Builder: ViewControllerBuilder
 
     /// Type, responsible for performing a transition.
     associatedtype Transition: ViewTransition